refactor: move to authentication package
