+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Coffeecounter">
         <activity
-            android:name=".AuthenticationActivity"
+            android:name=".ui.authentication.AuthenticationActivity"
             android:exported="false"
             android:theme="@style/Theme.Coffeecounter" />
         <activity
