+++ app/src/main/java/org/feature/fox/coffee_counter/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import dagger.hilt.android.AndroidEntryPoint
-import org.feature.fox.coffee_counter.ui.HomeView
+import org.feature.fox.coffee_counter.ui.authentication.HomeView
 import org.feature.fox.coffee_counter.ui.theme.CoffeeCounterTheme
 
 @AndroidEntryPoint
