+++ app/src/main/java/org/feature/fox/coffee_counter/ui/authentication/AuthenticationActivity.kt
@@ -1,11 +1,10 @@
-package org.feature.fox.coffee_counter
+package org.feature.fox.coffee_counter.ui.authentication
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import dagger.hilt.android.AndroidEntryPoint
-import org.feature.fox.coffee_counter.ui.AuthenticationView
 import org.feature.fox.coffee_counter.ui.theme.CoffeeCounterTheme
 
 @AndroidEntryPoint
